+++ navbridge/reporter/__init__.py
@@ -1,8 +1,12 @@
+from navbridge.reporter.audit_manifest import AuditManifest, build_audit_manifest, write_audit_manifest
 from navbridge.reporter.json_reporter import report_to_json, write_json_report
 from navbridge.reporter.markdown_reporter import report_to_markdown, write_markdown_report
 from navbridge.reporter.policy_advisor import recommend_tolerance_bps
 
 __all__ = [
+    "AuditManifest",
+    "build_audit_manifest",
+    "write_audit_manifest",
     "report_to_json",
     "write_json_report",
     "report_to_markdown",